Creade smaller card components for minion card
