+++ src/lib/components/card/card-minion-amount.svelte
@@ -1,0 +1,10 @@
+<script lang="ts">
+  import type { MinionSeller } from "$lib/types";
+  import { getContext } from "svelte";
+  const minion = getContext<MinionSeller>("minion");
+</script>
+
+<div class="relative inline-flex w-0 flex-1 items-center justify-center overflow-hidden rounded-br-lg text-sm font-medium text-neutral-200">
+  <span class="z-10 inline-block flex-shrink-0 rounded-full px-2 py-0.5 text-xs font-medium text-neutral-800 transition-transform duration-300 group-hover:scale-125 group-hover:text-neutral-900">{` Amount: ${minion.amount}`}</span>
+  <div class="absolute z-0 h-5 w-20 flex-shrink-0 rounded-[50px] bg-neutral-400 transition-all duration-500 group-hover:h-full group-hover:w-full group-hover:rounded-none" />
+</div>
