+++ src/lib/components/card/index.ts
@@ -1,0 +1,2 @@
+import Root from "./cardminion.svelte";
+export { Root as MinionCard };
